+++ fstickbackend/Fstick registry service/.gitattributes
