+++ fstickbackend/Fstick registry service/.gitignore
