+++ fstickbackend/Fstick registry service/build.gradle
