+++ fstickbackend/Fstick registry service/gradle/wrapper/gradle-wrapper.jar
