+++ fstickbackend/Fstick registry service/gradlew.bat
