+++ fstickbackend/Fstick registry service/settings.gradle
