+++ fstickbackend/Fstick registry service/src/main/java/ru/fstick/registry_service/FstickRegistryServiceApplication.java
