+++ fstickbackend/Fstick registry service/src/main/java/ru/fstick/registry_service/dto/api/request/PluginRequest.java
@@ -1,4 +1,4 @@
-package ru.fstick.registry_service.dto.api.requests;
+package ru.fstick.registry_service.dto.api.request;
 
 import jakarta.validation.constraints.*;
 import lombok.Data;
