+++ fstickbackend/Fstick registry service/src/main/java/ru/fstick/registry_service/dto/api/view/PluginsView.java
@@ -1,16 +1,16 @@
-package ru.fstick.registry_service.dto.api.responsies;
+package ru.fstick.registry_service.dto.api.view;
 
 
 import lombok.Builder;
 import lombok.Data;
-import ru.fstick.registry_service.dto.bd.PluginData;
+import ru.fstick.registry_service.dto.model.PluginData;
 import ru.fstick.registry_service.dto.service.PaginationData;
 
 import java.util.List;
 
 @Data
 @Builder
-public class PluginsResponse {
+public class PluginsView {
     private List<PluginData> items;
     private PaginationData pagination;
 }
