+++ fstickbackend/Fstick registry service/src/main/resources/application.properties
