+++ fstickbackend/Fstick registry service/src/test/java/ru/fstick/registry_service/FstickRegistryServiceApplicationTests.java
