+++ fstickbackend/Fstick registry service/swagger-api/registry service/openapi.yaml
